完善用户列表页面
完善一些用户列表页面的小细节，添加表格 loading 功能

+++ src/components/Users/index.js
@@ -13,19 +13,24 @@ export default {
         pagesize: 2,
         query: ''
       },
-      total: 10
+      total: 10,
+      userListLoading: true
     }
   },
   methods: {
     async getUserList () {
+      this.userListLoading = true
       const { data, meta: { status } } = await Users.getUserList(this.userListInfo)
       if (status === 200) {
         if (this.userListInfo.pagenum !== 1 && data.users.length === 0) {
           this.userListInfo.pagenum--
           return this.getUserList()
         }
-        this.userList = data.users
-        this.total = data.total
+        setTimeout(() => {
+          this.userList = data.users
+          this.total = data.total
+          this.userListLoading = false
+        }, 300)
       }
     },
     async changeState (id, state) {
@@ -35,12 +40,23 @@ export default {
         this.getUserList()
       }
     },
-    async userDestroy (uId) {
-      const { meta: { status, msg } } = await Users.destroyUser(uId)
-      if (status === 200) {
-        this.$message.success(msg)
-        this.getUserList()
-      }
+    userDestroy (uId, username) {
+      const ce = this.$createElement
+      this.$confirm(ce('p', null, [
+        ce('span', '次操作讲永久删除 '),
+        ce('h1', { style: 'display: inline;' }, username),
+        ce('span', ' 用户，是否继续？')
+      ]), '确认删除', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(async () => {
+        const { meta: { status, msg } } = await Users.destroyUser(uId)
+        if (status === 200) {
+          this.$message.success(msg)
+          this.getUserList()
+        }
+      }).catch(() => this.$message.info('已取消删除'))
     }
   },
   watch: {
